[DEV] modify "for Live Version" // DEV_96
Release 환경 최적화 ( Live Ver )
 4. Woker Thread 수 제어
 5. 코드 분석 피드백 적용

+++ Server/Server/IOCPServer/IOCPServer.cpp
@@ -244,9 +244,10 @@ void IOCPServer::_InitWinSocket()
 	// 일반적으로 스레드 풀의 크기는 프로세서 개수의 2배 정도를 할당한다.
 	// 하지만 과연 이 2배라는 수가 내가 만든 서버에서 가장 최적의 성능을 낼 수 있다고 보장할 수 있는 것일까..
 	HANDLE hThread;
-	//for (int i = 0; i < (int)si.dwNumberOfProcessors * 2; ++i)
 
-	for (int i = 0; i < (int)2; ++i)	// 현재 2 쓰레드만 사용.
+	// 쓰레드 수 제어.
+	//for (int i = 0; i < (int)si.dwNumberOfProcessors * 2; ++i)
+	for (int i = 0; i < 2; ++i)
 	{
 		hThread = CreateThread(NULL, 0, WorkerThread, (LPVOID)this, 0, NULL);
 		if (hThread == NULL)
